+++ tests/testthat/test-convert_param.R
@@ -6,9 +6,10 @@ cells_b <- c("B1_C1", "B2_C1", "B1_C2", "B2_C2")
 # single number ---
 testthat::test_that("single number", {
   param <- 3
-  cp <- convert_param(param, cells_w, cells_b) %>%
-    unlist() %>% magrittr::equals(param) %>% sum()
-  expect_equal(cp, 16)
+  cp <- faux:::convert_param(param, cells_w, cells_b)
+  comp <- matrix(rep(3, 16), 4, dimnames = list(cells_b, cells_w)) %>% as.data.frame()
+  
+  expect_equal(cp, comp)
 })
   
 # list of single numbers ----
@@ -19,16 +20,21 @@ testthat::test_that("list of single numbers", {
     "B1_C2" = 3, 
     "B2_C2" = 4
   )
-  cp <- convert_param(param, cells_w, cells_b) %>% unlist() %>% unname()
-  expect_equal(cp, rep(1:4, each = 4))
+  cp <- faux:::convert_param(param, cells_w, cells_b)
+  comp <- matrix(rep(1:4, times = 4), 4, dimnames = list(cells_b, cells_w)) %>% 
+    as.data.frame()
+  expect_equal(cp, cp)
 })
 
 
 # unnamed vector of cells ----
 testthat::test_that("unnamed vector of cells", {
   param <- rep(1:4, each = 4)
-  cp <- convert_param(param, cells_w, cells_b) %>% unlist() %>% unname()
-  expect_equal(cp, param)
+  cp <- faux:::convert_param(param, cells_w, cells_b)
+  comp <- matrix(rep(1:4, times = 4), 4, dimnames = list(cells_b, cells_w)) %>% 
+    as.data.frame()
+  
+  expect_equal(cp, comp)
 })
 
 
@@ -40,8 +46,9 @@ testthat::test_that("list of unnamed vectors", {
     "B1_C2" = 9:12,
     "B2_C2" = 13:16
   )
-  cp <- convert_param(param, cells_w, cells_b) %>% unlist() %>% unname()
-  expect_equal(cp, 1:16)
+  cp <- faux:::convert_param(param, cells_w, cells_b)
+  comp <- matrix(1:16, 4, dimnames = list(cells_w, cells_b)) %>% t() %>% as.data.frame()
+  expect_equal(cp, comp)
 })
 
 
@@ -53,8 +60,9 @@ testthat::test_that("list of named vectors", {
     "B1_C2" = c("W1_X1" = 9, "W2_X1" = 10, "W1_X2" = 11, "W2_X2" = 12),
     "B2_C2" = c("W1_X1" = 13, "W2_X1" = 14, "W1_X2" = 15, "W2_X2" = 16)
   )
-  cp <- convert_param(param, cells_w, cells_b) %>% unlist() %>% unname()
-  expect_equal(cp, 1:16)
+  cp <- faux:::convert_param(param, cells_w, cells_b)
+  comp <- matrix(1:16, 4, dimnames = list(cells_w, cells_b)) %>% t() %>% as.data.frame()
+  expect_equal(cp, comp)
 })
 
 # list of disordered named vectors ----
@@ -65,42 +73,42 @@ testthat::test_that("list of disordered named vectors", {
     "B2_C1" = c("W1_X1" = 5, "W1_X2" = 7, "W2_X1" = 6, "W2_X2" = 8),
     "B2_C2" = c("W1_X1" = 13, "W1_X2" = 15, "W2_X1" = 14, "W2_X2" = 16)
   )
-  cp <- convert_param(param, cells_w, cells_b) %>% unlist() %>% unname()
-  expect_equal(cp, 1:16)
+  cp <- faux:::convert_param(param, cells_w, cells_b)
+  comp <- matrix(1:16, 4, dimnames = list(cells_w, cells_b)) %>% t() %>% as.data.frame()
+  expect_equal(cp, comp)
 })
 
-
-# as data frame ----
-testthat::test_that("as data frame", {
+# data frame ----
+testthat::test_that("data frame", {
   param <- data.frame(
-    1:4,
-    5:8,
-    9:12,
-    13:16
-  ) %>%
-    magrittr::set_names(cells_b) %>%
-    magrittr::set_rownames(cells_w)
+    c(1, 5, 9, 13),
+    c(2, 6, 10, 14),
+    c(3, 7, 11, 15),
+    c(4, 8, 12, 16),
+    row.names = cells_b
+  )
+  names(param) <- cells_w
   
-  as.list(param) %>%  lapply(magrittr::set_names, rownames(param))
-  
-  cp <- convert_param(param, cells_w, cells_b) %>% unlist() %>% unname()
-  expect_equal(cp, 1:16)
-  
+  cp <- faux:::convert_param(param, cells_w, cells_b)
+  comp <- matrix(1:16, 4, dimnames = list(cells_w, cells_b)) %>% t() %>% as.data.frame()
+  expect_equal(cp, comp)
 })
 
 # backwards data frame ----
 testthat::test_that("backwards data frame", {
   param <- data.frame(
-    c(1, 5, 9, 13),
-    c(2, 6, 10, 14),
-    c(3, 7, 11, 15),
-    c(4, 8, 12, 16)
-  ) %>%
-    magrittr::set_names(cells_w) %>%
-    magrittr::set_rownames(cells_b)
+    1:4,
+    5:8,
+    9:12,
+    13:16,
+    row.names = cells_w
+  )
+  names(param) <- cells_b
   
-  cp <- convert_param(param, cells_w, cells_b) %>% unlist() %>% unname()
-  expect_equal(cp, 1:16)
-  
+  cp <- faux:::convert_param(param, cells_w, cells_b)
+  comp <- matrix(1:16, 4, dimnames = list(cells_w, cells_b)) %>% t() %>% as.data.frame()
+  expect_equal(cp, comp)
 })
+
+